+++ Assets/Sources/Components/ZoneComponent.cs
@@ -5,7 +5,7 @@ using Entitas.CodeGeneration.Attributes;
 [Game, Unique]
 public sealed class ZoneComponent : IComponent
 {
-    public int id;
+    public ushort id;
 
     public GameEntity GetZone(GameContext gameContext) => gameContext.GetEntityWithId(id);
 }
